requirements.txt
pandas
numpy
pyyaml
pillow
streamlit
xarray
h5netcdf
# optional
pygbif
